[sources] Don't block on parsing for execution line/token highlighting.
Previously we'd block the main thread for up to 500ms in the
DebuggerPlugin when loading a script in order to construct the
AST for the execution line and token highlighting (in particular
for the token actually).

With this CL we instead use a ViewPlugin for the execution location
highlighting (together with a StateField to store the position),
which allows us to work with the partial syntax tree without having
to block on parsing. CodeMirror will call `update()` on the plugin
as it progresses with incremental parsing.

Bug: 333945853, 40938371
Change-Id: I1758d769cc1737703a440502ec6f39cf7aef4702