[testing] Correctly pass chromeBinary to karma
Karma's chrome launcher respects the CHROME_BIN env var to configure the Chrome binary to launch, and launches the system's chrome if unset, which is currently the case. We don't want to pollute the environment vars for passing settings around, so we instead subclass the Chrome launcher to use the correct default chrome.

Bug: b:333423685
Change-Id: I546d25ad84e43ca76ee45734b4c69ccffb01f63a