[A11y] Update shortcut announcement for screen reader (for mac)
This patch updates this CL here - https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5292435 to fix screen reader announcements for shortcut changes across platforms.

The original patch used the `UI.ARIAUtils.alert` helper which seemed to work only for narrator on windows but not with voiceover on mac. This patch updates the strategy to set focus on an info element shown on shortcut changes in order to trigger the screen reader announcements.

 
Screenshot (click to play gif): https://imgur.com/a/gzwC63U

Bug: 324930007
Change-Id: I4c9009eef41bb61109bd5bc5bd65cf5c72057988