Add branch number property to branch builders
Test runners in dtf are getting migrated to a new unified runner. The devtools_frontend recipe needs to be able to remain capable of running tests with legacy runners on the older branches. To determine if we are in a legacy runner branch we will pass the branch number property to the
recipe.

Bug: 333423685
Change-Id: Id07fe81f096459455c5a0a2a3408dfdce680295f