[formatter] Simplify AcornTokenzier for JavaScript formatting
The parser already produces a token stream, we don't need to tokenize
a second time (it also uses different options for some reason).

This greatly simplifies the logic in the AcornTokenzier, as we don't
need to sync a comment stream with a token stream. We can just dump
everything into a single array and iterate that.

The performance gains are negligable, but the code becomes much
simpler.

Drive-by: Remove tokenColumnStart as well, it was unused.

R=bmeurer@chromium.org

Bug: 333945853
Change-Id: Iea283c012765842c37e824740dd52ddab380bfed