[sources] Use canonical script ID to identify UISourceCodes.
I noticed that with the repro from http://crbug.com/333945853 we don't
always reliably show a location line/token highlighting at all. This
was because we were checking for identity of UISourceCodes, but the
`blitz.<HASH>.js` file is provided by many different workers and there
are no guarantees that we always end up with the canonical version of
the UISourceCode.

Bug: 333945853, 40846859
Change-Id: Id13c13b7da124cfa43e99fc93051b7038cffba2e