RPP: expand ability to resolve nodeIds from events
The old trace engine has a few extra events that it can resolve nodeIds
from which we then use to give us the ability to highlight them, etc, in
the UI.

Before we can remove the old engine we need to add this functionality to
the panel via the new engine. This CL introduces a function that will do
just that, calculating the node ids for a given event. This is cached
too (future lookups will involve more work and thus we won't want to
re-run them for the same event).

Bug: 40287735
Change-Id: I8f4716c2613900b03ba53ed93160c28666c1b5ce