[ve] Remove unncessery observer instances from LoggingDriver and tests
Only using one instance per observer type in the implementation and
disconnecting them in stopLogging. Expect calls to the thottler instead
of various hacks to drain the queue in tests.

This brings ensures the tests robustness needed to the next CL where we
address the issues with the resize throttling implementation.

Bug: 331569913
Change-Id: Ibea4cc5617a4f5185dd61c7f86251d6080285f54