[Compression Dictionary Transport] Introduce SharedDictionaryIssue
Currently it is difficult to know the reason for the dictionary
write failure. To show the reason, this CL introduces a new
SharedDictionaryIssue class in DevTools frontend.

The SharedDictionaryIssue is sent from a NetworkServiceDevToolsObserver
in the browser process (https://crrev.com/c/5446494). And this
SharedDictionaryIssue class will show the issue in the Issue tab of
DevTools.

Bug: 333756098
Doc: go/chrome-devtools:shared-dictionary-issues-proposal
Demo: https://devtools-dbg-stories.netlify.app/crbug-333756098
After: https://crbug.com/333756098#comment3

Bug: 333756098
Change-Id: Iaf59c2729023c10ba21ca3cc1f0bc13100e48037