[RPP] Do not refresh tree view when not visible
TimelineTreeview.refreshTree was called in different views
(Bottom-Up and Call Tree) by the same single event when a group-by
dropdown changes value. Thus, we bail out whenever the view is  not
visible, which we know if the related element is detached  from the
document.

Failing to bail out might mean calculations done to refresh the trees of
different views can conflict with each other when they are scheduled to
the renderer coordinator.

Fixed: 325121014
Change-Id: Icb4fd1a48091f7222f85017c28ad3d27a4b1cd6c