[GM3Restyling] Use `ToolbarMenuButton` consistently.
Prior to this CL we had three occurences of `ToolbarButton`s that would
turn into selects with a menu attached to it, that would not use the
`ToolbarMenuButton`, but rather use a plain `ToolbarButton` with custom
menu logic. Part of the problem was that those three occurences would
also force a soft context menu (for no reason). These were:

1. The level selector in the Console panel.
2. The experimental request type selector in the Network panel.
3. The experimental 'More filters' selector in the Network panel.

With this CL we move the `turnIntoSelect()` method to the
`ToolbarMenuButton` class to enforce the invariant that all of
these select buttons in toolbars now consistently use the same
implementation (and the same menu logic and appearance).

As a consequence the above mentioned occurences now also use the
native OS menu by default (unless general usage of soft context
menu is enforced, for example for the purpose of E2E testing).

Bug: 330213829
Change-Id: I8c4bb26b284c48e4370cb340bb6bcca3a3df9e7a
Screenshot: https://i.imgur.com/1vVO3dt.png