RPP: don't use the old model in buildTraceEventDetails
This CL updates TimelineUIUtils.buildTraceEventDetails to not take the
old model as an argument (it was unused).

As part of this clean-up I also migrated the tests to not load all the
legacy models. That lead me to discover two issues:

1. We relied on the fact that TraceLoader.allModels also sets the trace
   bounds. I created a new helper to make this easy in future tests, and
   also made TraceLoader.traceEngine clear any old bounds to avoid tests
   depending on each other.
2. We rely on the fact that for some events, the old engine will
   subtract one from their line numbers to make them 0 indexed. I have a
   fix in mind for this where we do the same thing but in the UI layer
   (not mutating the input events). However that is a bigger change than
   I want in this CL, so I have ignored it for now and filed
   crbug.com/336734268 which I will work on next.

I also marked `TraceLoader.allModels` as deprecated as we should not be
using it for new tests and our goal is to get rid of it.

Bug: 40287735
Change-Id: I7a00cf5adb8a3c5ea39997a08aa39a7240b213e9