RPP: add backendNodeIds for DecodeImage
This CL adds the final resolution of backendNodeIds to the new engine,
for `DecodeImage` and `DrawLazyPixelRef`.

Note that I have not included ResizeImage as that event no longer exists
and we need to remove it from the panel (crrev.com/c/5481691).

In a follow-up CL I will remove `backendNodeIds` from
`EventOnTimelineData` and update all usages to rely on the new
FetchNodes helper only.

Fixed: 336237398
Change-Id: I2190770b2c12bffe2e8698a2fb8b7d588af0a1c7