[sources] Bail in decoration computation when the DebuggerPlugin died
The rule for the DebuggerPlugin is basically: Check `this.editor` after
every await. When a CodeMirror editor dies, we dispose of the
corresponding DebuggerPlugin. The plugin itself resets it's data
structures and nulls `this.editor`. But there might still be
continuations queued up in the micro task queue that will continue
in the disposed DebuggerPlugin.

Ideally we had something like cancellable promises/tasks, and cancel
all the in-flight tasks when a DebuggerPlugin dies.

R=bmeurer@chromium.org

Bug: 336857653
Change-Id: I3ee3eff933d3946d98396f7d3830d270bf4897ae