RPP: fix breadcrumbs test
This test assumed that a TraceBounds instance existed, but it might not
(unless some other test set one up). To fix, I ensure a new instance is
initialised with the right data in the test.

I've also added logging to catch if this happens in other tests so we
can easily fix in the future.

Bug: none
Change-Id: Icc3a16a08c29c01af8043e25f92637e817c001d3