Remove 'Consistent source map variable experiment'
The experiment has been enabled since M121 and UMA shows that it was
only disabled by a handful developers.

The CL also removes an e2e test that relied on the old behavior where
we would source map the range the users is hovering over and evaluate
that instead of doing variable substitution.

R=bmeurer@chromium.org

Fixed: 40944633
Change-Id: I2b9adc5e2f3ac71030c419e6a7248c93a2487a9e