RPP: Remove TimelineModel.PageFrame
This CL replaces it with a helper and replaces the usage in
TimelineTreeView that groups by frames. It also updates the test because
the old engine would leave sets of events unattributed, whereas now we
default to the main frame more readily (it looks like the old model
meant to do this but didn't).

This CL also removes the custom context menu in TimelineTreeView that
only applied when `pageFrame.ownerNode` was set. However, looking at the
PageFrame code relating to ownerNode, I found:

```
// TODO(dgozman): figure this out.
// this.ownerNode = target && payload['nodeId'] ? new SDK.DOMModel.DeferredDOMNode(target, payload['nodeId']) : null;
this.ownerNode = null;
```

And `ownerNode` was never set elsewhere in the codebase. Therefore I
have removed all that logic.

Fixed: 336283309
Bug: 40287735
Change-Id: I4e5ddfcd4823634464248794f2b4b21fd4654994