Trace: use isOutermostMainFrame to determine main frame ID
In crrev.com/c/5424783 we added `isOutermostMainFrame` to
TracingStartedInBrowser events to enable us to use this as a more robust
way to determine which frame we should call the main frame. This fixes
some instances where our previous heuristics would cause false
positives.

This Cl updates the MetaHandler to use this flag if it is available to
determine the main frame, but falls back to the previous approach if it
is not available. This ensures we don't break behaviour for any saved
trace files.

Fixed: 41494290
Change-Id: I0679aaec2edeaa3067c6e6d2864c651f7b00bcf6