RPP: move TreeView to new engine types
This CL updates the majority of the types in TimelineTreeView and
associated files to using the new engine only.

The one exception is the TimelineModel.forEachEvent helper, which for
now I want to leave alone, but other than that the tree views now use
the new events only.

This isn't a functional change, because all events are from the new
engine now, but it's a step towards removing the legacy implementation
and the CompatibleEvent type that has supported us through the
migration.

Bug: 40287735
Change-Id: I9de826b15d278306c9ee89487b0083c231741807