[ServiceWorker] Check optinal for ServiceWorkerRouterInfo.ruleIdMached
This CL adds check if `SerivceWorkerRouterInfo.ruleIdMatched` is optional
or not. In https://crrev.com/c/5481450, the field has been modifed to be
an optional number. To support this, we check if the field is optional
or not before accessing the field. If the field is optional, we pass on
0 which indicates an invalid rule ID.

Bug: 335524390
Change-Id: I053d32d9fed46d2bd37d81c8164d80a0e843385a