RPP: remove some usages of legacy models and events
This CL began as me updating the TimelineSelection class to not allow
old style events to be selected, but that created a bit of a path of
legacy events and removing them. I was also as a consequence able to
remove a bunch of references to the timeline model and performance model
and reduce how often we pass them around.

Bug: 40287735
Change-Id: I804b686179e0e8f70ed3c2e50f59b43776088fb5