Fix TimelineUIUtils test
Two CLs landed in a way where they both passed individually but the
first then broke the second. This CL fixes the tests.

There are also some outdated screenshots that will be updated. This CL
also removes tests for the GPU track in dark mode - we have other global
Perf panel dark mode tests, so we don't need to test this on an
individual track level.

Bug: none
No-Tree-Checks: true
Change-Id: I385ac4d8196926e0ad0f95ad52855f47e40f3b96