RPP: reset SelectorStatsView on a new trace
We had a bug where the selector stats view would never be updated when
the user imported/recorded a second trace, so it would always break.

This CL updates the code to reset the SelectorStatsView when it gets a
new traceParsedData object, ensuring that the view is not running on old
data.

Fixed: 338197915
Change-Id: I02100a60f7761fa03ba0a14eade80a2275bc8d21