Set "protocol" to network protocol
This replaces the (unused) "protocol" field with the network
protocol[1], instead of URL.protocol. This matches the "protocol"
field in CDP.

Also a drive-by removal of other unusued fields from the parsed URL.
This data can be derived from "url", and nothing uses it, so we're just
wasting some memory here.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/5512645
    (note, CL message incorrectly says priority, but it is about
     protocol)

Bug: 335099340
Change-Id: I7203d1b9b8d45415297db70e311014d1981b08dc