[RPP] Improve types for synthetic trace events created @ treify
`SyntheticTraceEntry` was used for trace events containing a `selfTime`.
This property is added to trace events when the run the `treify` step
in the Renderer & Samples handlers.
This type was problematic for two reasons:

1. Raw trace events could directly be typed as such despite not really
   being synthetic. This means that the `selfTime` was added directly to
   the source trace events, which we save when exporting a trace file.
2. Its name was too generic, it could mean any synthetic event, but
   that's not its purpose.

To improve this, this CL renames `SyntheticTraceEntry` to
`SyntheticTreifiedEntry`, which is now an union between the existing
SyntheticProfileCall and SyntheticRendererEvent. The latter is also
changed to be a branded type so that no raw trace event can be typed as
such, hinting the developer to create a copy to prevent mutating the raw
trace data.

Fixed: 330204386
Change-Id: I281cd3e4442a926ed875b17befbc15315ebd23b8