RPP: replace TraceLoader.allEngines usage
In anticipation of the old engine being removed, we don't need to use it
in tests. We are also only using the new engine most of the time.

Note that the final usages of this method are in
TimelineFilters.test.ts, which currently do support both engines. In a
separate CL I will update them to support only the new engine and then
remove the usage of TraceLoader.allEngines.

Bug: 40287735
Change-Id: Ib24c3057f31f78d9a1ae130ce9760665c506394e