[test] Mock getResourceTree response by default.
Wheneve a target is created, ResourceTreeModel makes this call which
never returns. This results in frame targets without a
ResourceTreeFrame, which is very different from what happens in
production.

Also introduce a helper to wait this call and exempt a few tests that
heaviliy rely on targets without frames.

The next CLs will introduce more canonical helpers to mock various
resource tree states.

Bug: none
Change-Id: I5f1c9ecc3ed66d9936cbf96562f2276330a2498b