Remove CPU filter from TimelineLoader
This is not needed any more: the record types here are not used in the
new engine. Because the CPU Profile "mode" constructs tracks from
threads via the samples handler, all the "filtering" of events is done
for us, and we don't need this extra filter as it's a no-op anyway.

Bug: 40287735
Change-Id: Ib4974775e6c341bf2b078904f6f5c39459dbb4de