RPP: remove unused TimelineModel methods + properties
This CL removes a lot of code from TimelineModel. All of the removals
either had no references, or were instance properties that were created
and added to, but never used for any logic.

This file is close to being entirely removed, but going slowly bit by
bit like this gives us more chance to catch any problems as we go, hence
this CL.

Bug: 40287735
Change-Id: If4eb71195c4be55eb3aeeb8a000c42ea89c998ec