[ve] Approximate VE logging of overlay paused message
The actual overlay UI is in the separate document that cannot
communicate with DevTools frontend or UI bindings where we do the
logging. The buttons calls methods directly on the V8 session, so
out best bet is to guess what happened by listening to
Debugger.pasued/resumed and assume that if the frontend document had no
focus, then the action was triggered through the overlay.

Bug: 330714235
Change-Id: I16401c8babf595fafe860d5ec78cd99eb7de326e