Reland "Copy goldens in swarming context with new test runner"
This is a reland of commit 16010bc2d81d9a07dd68beaa5cf85276962dabf0

Pls note that the original CL did not close the tree. The build referenced in the revert is a led test run. The builder appeared in Milo  as a result of faulty Milo config that got fixed.

Original change's description:
> Copy goldens in swarming context with new test runner
>
> Bug: 333423685
> Change-Id: I21e37494646501cac83d6ab379e968c2e4f5d068
> Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5518457
> Reviewed-by: Andres Olivares <andoli@chromium.org>
> Commit-Queue: Liviu Rau <liviurau@google.com>

Bug: 333423685
Change-Id: If98a347781b6bd97cc2689b47687be011817e099