Make ThemeSupport self-contained
This CL proposes that ThemeSupport should own all the listener
registration/unregistration. The private constructor installs all the
necessary listeners on the DOM and the host bindings and in the
singleton 'instance' method we check dispose off the old instance if
needed.

The result is that calling ThemeSupport.instance({force: new}); always
results in a sane state.

To make this happen we turn `fetchColors` into an instance method.
It calls `applyTheme` in any case so this makes it more explicit that
`fetchColors` actually relies on the current global instance.

Bug: None
Change-Id: I12e80362d53838a281724db5d2e598523ac57cfe