RPP: disable the old engine entirely
This CL stops us calling the old TracingModel entirely and stops us
using both TracingModel and PerformanceModel (which is now removed from
the codebase).

This CL also opens us up to deleting more code (LegacyTracingModel is
also unused ), but I will do that in follow-up CLs as this one is big
enough already.

Once that code is gone, we can also remove any tests that test the
legacy engine and remove the helpers that construct the EventPayload
type.

Bug: 40287735
Change-Id: I58688eae8b8bab13cd080017907a9beae4fd5d6b