[infra] Use recursedeps for first-class GCS deps in build and buildtools
Currently, devtools downloads GCS dependencies via hooks. This requires
a sha1 hashsum in the dependent repository.

With the introduction of first class GCS dependencies, these hashsum
files will be deprecated in favor of managing GCS dependencies in the
deps section of the DEPS file.

This change adopts this migration by removing the artifact download in
the hooks section, and by adding the affected dependencies to the
recursedeps section.

This is currently supported for sysroot
(build, https://crrev.com/c/5519633) and clang-format
(buildtools, https://crrev.com/c/5506857) related artifacts.

Bug: 336898179
Change-Id: I04308df89fa6f6e10d54d24fb49cf002f16ea921