Don't list zero-size InternalNodes in heap snapshot summary
We used to need a special case to ensure that native nodes showed up in
the heap snapshot even though their size was reported as zero. However,
since https://crrev.com/c/3021169 , the sizes for named native nodes are
now reported. There is no need to list the remaining unnamed native
nodes in the Summary view.

Bug: 340200025
Change-Id: I3a2ea90455caf7df2e06c34595921e27c7be1489