Fix and simplify theming around device emulation toolbar
The 'device_mode_emulation_frame' entrypoint is created via window.open,
as such, it shares a context with the DevTools window but gets its own
`window` object. Since the context is shared, we only create one
`host.js` ES module, which in turn only installs the
`window.FrontendHostAPI` on the DevTools `window` instance.

Even thoug the device mode emulation frame window has it's own
DevToolsUIBindings on the backend side, events from it can't arrive
since we never set up the proper infra on the emulation `window`. This
is not a problem, since it doesn't actually require host bindings, but
we would see 'TypeErrors' for non existing 'window.FrontendHostAPI'
objects non-theless in the console.

This CL simply checks for the existence and doesn't try to emit host
binding events if no FrontendHostAPI is found.

Furthermore, this CL simplifies 'ThemeSupport' w.r.t. to multiple
documents: Since we share a context between the DevTools window and
the device emulation toolbar window, we also share a 'ThemeSupport'
instance. So instead of doing all the manual invoking and wiring in
AdvancedApp.ts, it is enough to add support for multiple `Document`
instances to ThemeSupport.

This means that if we get a 'theme changed' event from the backend,
or if the dark/light mode setting changes, we install the 'style' tag
and re-apply the CSS variables to all known documents.

Drive-by: We now unconditionally emit the 'ThemeChanged' event, not
only when switch between dark/light mode just to keep semantics
simpler.

R=jacktfranklin@chromium.org, kprokopenko@chromium.org

Bug: None
Change-Id: I10c829beee2869bd999d7941a636ee46f9534183