[sdk] Implement decoding of original scope references
This CL adds decoding of "definitions" of generated ranges. A
definition is a source file index plus a scope index where the
scope index references the index of the corresponding "start" item.

For now, we flatten out the resulting scope tree manually and insert
'null' for end items. If this becomes performance critical we can
construct an array or a map (index -> OriginalScope) as
we decode the original scopes themselves.

R=pfaffe@chromium.org

Bug: 40277685
Change-Id: I5a86c1fdfc7af330fce3d37f2854c9fa734aac78