RPP: URLForEntry helper
This CL merges some duplicate functions we had that, for a given entry,
can return the URL.

This has to live in Extras rather than Helpers because Handlers depends
on Helpers, so you cannot have Helpers depending on Handlers too. That's
a bit frustrating but...Extras seems like a decent place for these. We
can think about this some other time.

Bug: 40287735
Change-Id: I76cce3025864e24b14d7d7c3bdc8035ed551969d