[RPP] Rename "modified" entries to expandable
Entries whose children have been collapsed are called "modified"
throughout the modifications artifacts, meaning they contain an arrow
showing they can be "expanded" to reveal the hidden entries.

Because "modified" is a wide term and we intend to support modifications
beyond the expandable mark, this CL changes the "modified" adjective to
"expandable" to be more specific.

Bug: 339163152
Change-Id: Ia1757310e3f092e3213cee33220ac8e44f787cf9