[persistence] Simplify file reading
This CL uses the newer Blob.text() method to read file contents from
disk.

This CL also changes the logic to determine whether a file has binary
or text content. First off, the Blob has a mime type itself attached,
so Chrome might actually already tell us what kind of file we are
dealing with.

Afterwards we check the extension for a known mime type and then
we check the list of known binary extensions. If none of it hits,
we assume text content.

R=pfaffe@chromium.org

Bug: 324929074
Change-Id: I24eb5e6c989bf9b96d16a8aad646933354752495