[RPP] Add live metrics experiment to the performance panel
This uses the web-vitals library to install live metric listeners on the
target page and surface them on the performance panel landing page.

Other supplemental data is included such as the LCP element node.

The initial UI is extremely simple but we will continue to iterate on
that behind the experiment.

Bug: 313906462
Change-Id: Idb6aaaa1a0d5f542a1767218e3f8f4483d010658